move metrics and loss to their own modules
