+++ composer/loss/__init__.py
@@ -1,0 +1,1 @@
+from composer.loss.loss import soft_cross_entropy as soft_cross_entropy
