+++ composer/metrics/__init__.py
@@ -1,0 +1,9 @@
+"""A collection of common torchmetrics."""
+
+from composer.metrics.metrics import CrossEntropyLoss, Dice, LossMetric, MIoU
+from composer.metrics.nlp import BinaryF1Score, HFLanguageCrossEntropyLoss, LanguageCrossEntropyLoss, MaskedAccuracy, Perplexity
+
+__all__ = ["MIoU", "Dice", "CrossEntropyLoss", "LossMetric",
+           "Perplexity", "BinaryF1Score", "HFLanguageCrossEntropyLoss",
+           "LanguageCrossEntropyLoss", "MaskedAccuracy"]
+
