+++ composer/models/deeplabv3/deeplabv3.py
@@ -11,9 +11,9 @@ from torchmetrics import MetricCollection
 from torchvision.models import _utils, resnet
 
 from composer.core.types import BatchPair
-from composer.models.base import ComposerModel
-from composer.metrics.metrics import CrossEntropyLoss, MIoU
 from composer.loss.loss import soft_cross_entropy
+from composer.metrics.metrics import CrossEntropyLoss, MIoU
+from composer.models.base import ComposerModel
 from composer.models.model_hparams import Initializer
 
 __all__ = ["deeplabv3_builder", "ComposerDeepLabV3"]
