+++ composer/models/unet/unet.py
@@ -11,8 +11,8 @@ import torch.nn as nn
 from torchmetrics import Metric, MetricCollection
 
 from composer.core.types import BatchPair
-from composer.models.base import ComposerModel
 from composer.metrics.metrics import Dice
+from composer.models.base import ComposerModel
 from composer.models.unet.model import UNet as UNetModel
 
 log = logging.getLogger(__name__)
